Finished the specials, started on the Items
The specials are officially finished, probably will need to twik them a little in the future but for now I see no problem on them.
Started on the items, still need to put comments but today was an exhausting day and will call it now. Great work even though I'm in Esmoriz and got my mom constantly distracting me, in a good way, but still distracting me.